# shared/openapi.yaml
openapi: 3.1.0
info:
  title: cinghialapp API
  version: 0.1.0
  description: |
    D&D 5e campaign manager API.
    Server-authoritative dice. Role-based visibility (player/master).
servers:
  - url: http://localhost:8080/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /campaigns/{id}/sessions:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:  { tags: [recap], responses: { '200': { description: list } } }
    post: { tags: [recap], responses: { '201': { description: created } } }
  /sessions/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:    { tags: [recap], responses: { '200': { description: ok } } }
    patch:  { tags: [recap], responses: { '200': { description: ok } } }
    delete: { tags: [recap], responses: { '204': { description: deleted } } }
...
  /spells:
    get:
      tags: [spells]
      parameters:
        - { in: query, name: q, schema: { type: string } }
        - { in: query, name: level, schema: { type: integer } }
        - { in: query, name: class, schema: { type: string } }
      responses: { '200': { description: list, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Spell' } } } } } }
  /spells/{slug}:
    get:
      tags: [spells]
      parameters: [{ in: path, name: slug, required: true, schema: { type: string } }]
      responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Spell' } } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    Spell:
      type: object
      required: [slug, name, level, school, description]
      properties:
        slug: { type: string }
        name: { type: string }
        level: { type: integer, minimum: 0, maximum: 9 }
        school: { type: string }
        casting_time: { type: string }
        range_text: { type: string }
        components: { type: string }
        duration: { type: string }
        classes: { type: array, items: { type: string } }
        ritual: { type: boolean }
        concentration: { type: boolean }
        description: { type: string }
        higher_levels: { type: string, nullable: true }
        source: { type: string }
...